Animated model rendering.

+++ Insight/inc/Engine/Graphics/GPUBuffer.h
@@ -11,7 +11,7 @@ struct Vertex
 	glm::vec3 Normal;
 	glm::vec4 Colour;
 	glm::vec2 UV1;
-	glm::vec4 JointIndices;
+	glm::ivec4 JointIndices;
 	glm::vec4 JointWeight;
 #ifdef IS_MESH_BATCHING_EXT
 	u32 VIndex;
